+++ rubato/classes/components/image.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Dict
 import sdl2, sdl2.ext, sdl2.sdlgfx, sdl2.surface, sdl2.sdlimage
 
 from . import Component, Rectangle
-from ... import Vector, Display, Draw, Radio, get_path
+from ... import Vector, Display, Radio, Sprite, Draw
 
 if TYPE_CHECKING:
     from .. import Camera
@@ -29,10 +29,9 @@ class Image(Component):
 
     def __init__(
         self,
+        rel_path: str,
         offset: Vector = Vector(0, 0),
         rot_offset: float = 0,
-        rel_path: str = "",
-        size: Vector = Vector(32, 32),
         scale: Vector = Vector(1, 1),
         anti_aliasing: bool = False,
         flipx: bool = False,
@@ -42,21 +41,9 @@ class Image(Component):
         super().__init__(offset=offset, rot_offset=rot_offset, z_index=z_index)
 
         if rel_path == "":
-            self._image: sdl2.SDL_Surface = sdl2.SDL_CreateRGBSurfaceWithFormat(
-                0,
-                size.x,
-                size.y,
-                32,
-                sdl2.SDL_PIXELFORMAT_RGBA8888,
-            ).contents
-        else:
-            try:
-                self._image: sdl2.SDL_Surface = sdl2.ext.load_img(rel_path, False)
-            except OSError:
-                self._image = sdl2.ext.load_img(get_path(rel_path), False)
-            except sdl2.ext.SDLError as e:
-                fname = rel_path.replace("\\", "/").split("/")[-1]
-                raise TypeError(f"{fname} is not a valid image file") from e
+            raise ValueError("Image rel_path cannot be an empty string.")
+
+        self._sprite = Sprite(rel_path)
 
         self.singular = False
 
@@ -67,25 +54,23 @@ class Image(Component):
         self._resize_scale: Vector = Vector(1, 1)  # This scale factor is changed when the image is resized.
         self._rot = self.rotation_offset
 
-        self._original = Display.clone_surface(self._image)
-        self._tx = sdl2.ext.Texture(Display.renderer, self.image)
+        self._original = Display.clone_surface(self._sprite.image)
 
-        self._changed = True
-        self._update_rotozoom()
         self._go_rotation = 0
+        self._changed = True
 
         Radio.listen("ZOOM", self.cam_update)
 
     @property
     def image(self) -> sdl2.SDL_Surface:
         """The SDL Surface of the image."""
-        return self._image
+        return self._sprite.image
 
     @image.setter
     def image(self, new: sdl2.SDL_Surface):
-        self._image = sdl2.SDL_ConvertSurfaceFormat(new, sdl2.SDL_PIXELFORMAT_RGBA8888, 0).contents
-        self._original = Display.clone_surface(self._image)
-        self._update_rotozoom()
+        self._sprite.image = new
+        self._original = Display.clone_surface(new)
+        self._changed = True
 
     @property
     def scale(self) -> Vector:
@@ -153,9 +138,7 @@ class Image(Component):
         Returns:
             The size of the image
         """
-        if self.image.w == self._original.w and self.image.h == self._original.h:
-            return Vector(self.image.w, self.image.h) * self.scale
-        return Vector(self.image.w, self.image.h)
+        return self._sprite.get_size()
 
     def get_size_original(self) -> Vector:
         """
@@ -164,21 +147,16 @@ class Image(Component):
         Returns:
             Vector: The original size of the image.
         """
-        return Vector(self._original.w, self._original.h)
+        return self._sprite.get_size_original()
 
-    def _update_rotozoom(self):
-        """Updates the image surface. Called automatically when image scale or rotation are updated"""
+    def _update_sprite(self):
         if self.gameobj:
-            self._image = sdl2.sdlgfx.rotozoomSurfaceXY(
-                self._original,
-                -self.gameobj.rotation - self.rotation_offset,
-                # It seems that rotation is counterclockwise, even though we assume clockwise until now.
-                # Requires further investigation but is a fix for now.
+            self._sprite.rotation = -self.gameobj.rotation - self.rotation_offset
+            self._sprite.aa = self.aa
+            self._sprite.scale = Vector(
                 (-self.scale.x if self.flipx else self.scale.x) * self._resize_scale.x,
-                (-self.scale.y if self.flipy else self.scale.y) * self._resize_scale.y,
-                int(self.aa),
-            ).contents
-            self._tx = sdl2.ext.Texture(Display.renderer, self.image)
+                (-self.scale.y if self.flipy else self.scale.y) * self._resize_scale.y
+            )
 
     def resize(self, new_size: Vector):
         """
@@ -207,27 +185,22 @@ class Image(Component):
         self.resize(new_size)
 
     def draw(self, camera: Camera):
-        if self.hidden: return
+        if self.hidden:
+            return
 
         if self._changed or self._go_rotation != self.gameobj.rotation:
             self._go_rotation = self.gameobj.rotation
             self._changed = False
-            self._update_rotozoom()
+            self._update_sprite()
 
-        Draw.push(
-            self.true_z,
-            lambda: Display.update(
-                self._tx, camera.transform(self.gameobj.pos + self.offset - Vector(*self._tx.size) / 2))
+        Draw.sprite(
+            self._sprite, camera.transform(self.gameobj.pos + self.offset - Vector(*self._sprite.tx.size) / 2),
+            self.true_z
         )
 
     def delete(self):
         """Deletes the image component"""
-        self._tx.destroy()
-        sdl2.SDL_FreeSurface(self._image)
-        sdl2.SDL_FreeSurface(self._original)
-        self._image = None
-        self._tx = None
-        self._original = None
+        self._sprite.delete()
 
     def clone(self) -> Image:
         """
@@ -237,6 +210,7 @@ class Image(Component):
             Image: The cloned image.
         """
         new = Image(
+            "",
             offset=self.offset,
             scale=self.scale,
             anti_aliasing=self.aa,
@@ -248,42 +222,3 @@ class Image(Component):
         new.image = Display.clone_surface(self.image)
         new.rotation_offset = self.rotation_offset
         return new
-
-    @staticmethod
-    def from_surface(surface: sdl2.surface.SDL_Surface) -> "Image":
-        """
-        Creates an image from an SDL surface.
-
-        Args:
-            surface: the surface to create the image from.
-
-        Returns:
-            The created image.
-        """
-        # untested
-        image = Image()
-        image.image = surface
-        return image
-
-    @staticmethod
-    def from_buffer(buffer: bytes) -> "Image":
-        """
-        Creates an image from a buffer.
-
-        Args:
-            buffer: bytes containing the image data.
-
-        Returns:
-            The image created from the buffer.
-        """
-        # untested
-        rw = sdl2.SDL_RWFromMem(buffer, len(buffer))
-        surface_temp = sdl2.sdlimage.IMG_Load_RW(rw, 1)
-
-        if surface_temp is None:
-            raise Exception(sdl2.sdlimage.IMG_GetError())
-
-        surface = sdl2.SDL_ConvertSurfaceFormat(surface_temp, sdl2.SDL_PIXELFORMAT_RGBA8888, 0).contents
-        sdl2.SDL_FreeSurface(surface_temp)
-
-        return Image.from_SDL_Surface(surface)