+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 -   true_z property for components which gets the z_index of the `Component` offset by its parent `GameObject`
 -   `Time.scheduled_call()` for a self-correcting recurring method call on a timed interval.
 -   clone functions for `Group`, `GameObject`, and `Scene`.
+-   `Draw.texture` and `Draw.immediate_texture` to draw textures to the renderer
+-   `Draw.sprite` and `Draw.immediate_sprite` to draw sprites to the renderer
+-   `Sprite` class to draw images that are not linked to Game Objects
 -   `wrap()` function can create and populate a GameObject with component(s) automatically.
 
 ### Changed